refactor: HeaderBar 공통 컴포넌트 modules 폴더로 이동
